wee63: bug fixed
git-svn-id: http://grubutils.googlecode.com/svn@62 d95a8657-0e37-07c3-0af6-720afd5d0260

+++ grubutils/wee/builtins.c
@@ -1224,7 +1224,7 @@ map_func (char *arg/*, int flags*/)
   else
   {
     (*(unsigned short *)(void *)(&(bios_drive_map[i]))) = from | (to << 8);
-    grub_memset ((void *)(&(bios_drive_map[i])) + 2, 0, DRIVE_MAP_SLOT_SIZE - 2);
+    grub_memset ((void *)((int)(&(bios_drive_map[i])) + 2), 0, DRIVE_MAP_SLOT_SIZE - 2);
   }
   
   return 0;
